+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 # Version 0.1.0
 _2024-10-23_
 
+* Add `apollo-execution-federation` with minimal `@key` support
+* Add support for suspending resolvers and parallel execution (#18)
+* Update Ktor to version 3
+
 # Version 0.0.3
 _2024-08-30_